projects/ocr_label_extractor/ocr_label_extractor/requirements.txt
streamlit
opencv-python
pillow
google-generativeai
python-dotenv
pandas
openpyxl
gunicorn
requests

